requirements.txt
ultralytics
sentence-transformers
opencv-python
pillow
tqdm
rich
sqlite-vec
rembg
onnxruntime
pysqlite3-binary